requirements.txt
asgiref==3.5.2
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
python-decouple==3.6
Django==4.1
djangorestframework==3.13.1
djongo==1.3.6
drf-yasg==1.21.3
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
markdown==3.4.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pyparsing==3.0.9
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
pymongo==3.12.1
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.11
